+++ projects/map-maplibre/src/public-api.ts
@@ -2,6 +2,8 @@
  * Public API Surface of map-maplibre
  */
 
+export * from './lib/maplibre.helpers';
+export * from './lib/maplibre-layers.helpers';
 export * from './lib/map-maplibre.service';
 export * from './lib/map-maplibre.component';
 export * from './lib/map-maplibre.module';